backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.2.0
python-dotenv==1.0.0